refactor(contained): split the CLI front door and break an import cycle
Sentrux reported this branch as DEGRADED against main: quality 4820 ->
4463, god files 2 -> 3, complex functions 39 -> 43. Measuring rather than
assuming corrected the diagnosis twice.

**The god file was `cli/contained.py`, not `k8s_setup.py`.** Truncating
`k8s_setup.py` to sixty lines leaves the god-file count unchanged; the
score is lines weighted by dependency count, not lines. So `k8s_setup.py`
is untouched — splitting it would have cost a forty-site test-mock rewrite
and bought nothing measurable.

`cli/contained.py` goes 813 -> 167 lines, split into peers rather than one
file that did everything: `contained_args.py` reads the command line, and
`contained_local.py` runs one podman container — which makes the local
path a peer of `contained_k8s.py`, the shape the code already had
conceptually. `interpret` drops from cc 23 to under 13.

Five more functions come under the threshold by extracting what they were
each doing twice: the kubeconfig JSON-shape guard in `k8s.py`, the check
constructors in `verify_k8s`, the context list rendering, and the line
editor in `style.read_line`.

**The cycles were the whole story.** Everything above is worth five points.
Repointing two lazy imports in `contained_k8s.py` at `contained_args` —
deleting a shim that existed only to keep the old name working — removes
the `contained <-> contained_k8s` cycle and is worth 165. Quality is now
4637 of main's 4820, with god files and complex functions back at main's
numbers.

Quality is the geometric mean of five sub-scores, so the smallest one
dominates: acyclicity scores 2000 against redundancy's 9384. One cycle
remains, `k8s <-> k8s_division <-> lifecycle`, and it is a single strongly
connected component — breaking either edge alone measures no change. It
needs `sidecar_command` and the lifecycle types moved to modules they can
both depend on, which is a wider change than this one.
